WIP: h3: implement recv_body() API
The idea is to change the Data event to only return the amount of data
received, and then let the application fetch it from the transport
layer by using the new recv_body() method.

DATA frames handling is changed to only parse the frame header and leave
the payload in the transport buffer, and simply track how much data is left
to read for the current frame.

TODO:
 * Reset stream state after data is read completely.
 * Stream buffer handling needs to somehow be changed to only read the frame
   header from transport.
 * Handle DATA frame split across packets (e.g. generate new Data event).
 * Update / fix tests.
 * Add request / response with body tests.
 * Add HTTP/3 request / response with body benchmark.